setup hilt added use case

+++ base/src/main/java/ny/times/reader/base/presentation/ui/Chip.kt
@@ -47,4 +47,4 @@ fun Chip(
     }
 }
 
-data class ChipContent(val id: Int, val text: String, val isSelected: Boolean)
+data class ChipContent(val id: Long, val text: String, val isSelected: Boolean)
